Refactor packge.json scripts. Create production env config file.

+++ udemy-tdd-with-nodejs/test-cleanup.js
@@ -5,6 +5,10 @@ const config = require('config');
 const { uploadDir, profileDir } = config;
 const profileFolder = path.join('.', uploadDir, profileDir);
 
+if (!fs.existsSync(profileFolder)) {
+  return;
+}
+
 const files = fs.readdirSync(profileFolder);
 for (const file of files) {
   fs.unlinkSync(path.join(profileFolder, file));
